+++ modules/tinymce/src/core/main/ts/focus/EditorFocus.ts
@@ -119,6 +119,11 @@ const focusEditor = (editor: Editor) => {
     normalizeSelection(editor, rng);
   }
 
+  // This should be safe as the the focusin listener calls editor.focus(true) which means the focusEditor function is not called
+  // TODO: Should this also include the last blurredEditor?
+  // TODO: Need some tests for this
+  editor.fire('focus');
+
   activateEditor(editor);
 };